requirements.txt
Flask
Flask-Session
google-generativeai
